Introduced new "VimeoAccountType" enum values
We've encountered a user with "custom" as the account type, although this value isn't mentioned in Vimeo's own documentation of possible values. Prior to this commit, unrecognized values would trigger an exception, thus being unable to fetch affected videos/user.

To avoid a similar issues going forward, if the "account" omitted or empty, the type will be parsed to "Unspecified", whereas an unrecognized value will be parsed to "Unrecognized".